Restructure with internal implementation
